backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goal_bot"
version = "0.1.0"
requires-python = ">=3.11"
description = "Backend for WhatsApp goal tracking bot"
authors = [{name = "Your Name"}]
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
]

# Package discovery configuration
# Exclude 'db' directory to avoid setuptools error about multiple top-level packages
# Only include 'app' directory as the main application package
[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["db*"]
